PangoWin32: Add API to load custom font file
This adds an API for PangoWin32 to add a font file that is supported by
DirectWrite into the PangoFontMap, for Windows 10 and later.

If DirectWrite is able to load and gather the info on the font file, put it
into the IDWriteFontSetBuilder[1] member of the fontmap, and then inform
using pango_font_map_changed() which will call the private API in the previous
commit, which will cause the font map to be reloaded with the system font
collection plus the font file(s) that were passed in via this API, with the
needed signals and notifications sent.

The fontmap->changed() vfuncs for PangoWin32/PangoCairo(Win32) is updated
accordingly, for this reason.

Support for Windows 7/8.x would probably be added later, pending demand
for such support.